// src/app/pages/login/postulante/postulante-interface.ts
export interface PostulanteUpdate    {
    nombreUsuario: string;
    apellidoUsuario: string;
    ciudadUsuario: string;
    tipodocumentoUsuario: string;
    numerodocumentoUsuario: string;
    generoUsuario: string;
    direccionUsuario: string;
    descripcionUsuario: string;
    telefonoUsuario: string;
}